Split server/client libs
